+++ src/views/Dashboard/Dashboard.jsx
@@ -5,7 +5,7 @@ import BillDisplay from "../BillDisplay";
 import "./styles.scss";
 
 export default (props) => {
-
+  //testing with @testing-library to be learnt later
   const [activeState, setActiveState] = useState(
                                                 {
                                                   ...DEFAULT_ACTIVE_STATE,
@@ -14,12 +14,23 @@ export default (props) => {
                                               );
   const [component, setComponent] = useState(<BillDisplay />);
 
+  const filterComponentFromNav = (navType) => {
+    const filterKeyVal = {
+      "bills" : <BillDisplay/>,
+      "profile" : <div> Profile </div>,
+      "contacts": <div> Contacts </div>,
+      "settings": <div> Settings </div>,
+      "help": <div> Help </div>
+    };
+    setComponent(filterKeyVal[navType]);
+  };
+  
   return (
     <div>
       {localStorage.getItem("billSnap_token") ? (
         <div className="dashboard__flexbox">
           <div className="side__bar">
-            <Sidebar activeState={activeState} setActiveState={setActiveState} setComponent={setComponent}/>
+            <Sidebar activeState={activeState} setActiveState={setActiveState} filterComponentFromNav={filterComponentFromNav}/>
           </div>
           <div className="dashboard__content">
               {component}
